Add: gst-variable-rtsp-server application
This app already had support in the Makefile, but is now finally added. It
will be further documented in an included README, but the gist is that it
takes a video input (via v4l2src) and creates an RTSP server stream out of it.

It then changes bitrate of the video stream according to the number of clients
currently connected to it.

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>