src: gst-variable-rtsp-server: add dynamic bitrate changing (CBR)
Now that gstreamer-imx support changing bitrates on the fly in the
imxvpuenc_h264 encoder, add support for it. Also change to this behavior by
default. If bitrate is set to 0, it will use VBR instead (old behavior).

Also add support for setting the idr interval rate + some code cleanup. Finally,
bumped version to 1.1.

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>