remove source config and unecessary launch options
With the user passing the pipeline directly, there is no longer a need
to manually configure the source. There is an assumption that the user
will properly configure the source when using it in a pipeline.

The launch options 'i', 's', and 'f' have also been removed. The user
will manually specify the source and, if necessary, the device. Caps
filters can be passed into the pipeline, removing the need for the
launch option.