streamline pipeline setup and modification
Originally the size of the pipeline was hardcoded and relied on an array
of elements to control. By removing the array and just searching for
specific elements in the pipeline, there is no longer a need to hardcode
the pipeline size. This allows for the user to use a pipeline of any
size.

Signed-off-by: Matthew Michilot <mmichilot@gateworks.com>